Be more concrete about blogs
Blogs are now messages with an extra flag in their content to indicate that they are blogs. They will also look more like blogs to typical matrix clients.

Syncing is now better with presence totally ignored, and a large limit on the number of messages to be received: 100.

Fixed comment boxes