Simple Guest Access support
Add a button to login as a guest. This will register a guest for the user so that they can read blogs without registering properly.

TODO: blogs need to 'world readable' by default and more UI is needed for setting that but also displaying the status of the history readability. If a guest trys to view a blog that does not allow guest access, an error should be displayed, asking them to login.

BIG-TODO: allow guests to register.