Improve blog creation
Use SimpleMDE and marked to much improve the process of writing a new blog. This WYSIWYG editor can be used to send arbitrary HTML towards matrix, with support for headers, lists, formatted code, quotes, images, etc.

CSS modifications required for restricting images now embedded in content.

The editor now appears at the top of the page when a button is clicked and can be made full screen, with a live preview.